+++ CHANGELOG.md
@@ -734,6 +734,16 @@ of `NULL` was not set and read. (https://github.com/ClickHouse/clickhouse-java/i
   and `getObject(column, Object.class)` and the no-type `getObject(column)` overloads now return a decoded `String`
   instead of the internal holder.
 
+## 0.9.9
+
+### Bug Fixes 
+
+- **[client-v2]** `ServerException` with code `159 Execution Timeout` is retried unconditionally. After the fix this
+error treated as non-retriable. (part of https://github.com/ClickHouse/clickhouse-java/issues/2637)
+- **[jdbc-v2]** Fixes `Statement#setQueryTimeout`. By default, client executes query in calling thread and future timeout 
+has no effect. Fix makes `setQueryTimeout` to set `max_execution_time` server setting in this case to overcome limitation.
+  (https://github.com/ClickHouse/clickhouse-java/issues/2637)
+
 ## 0.9.8
 
 ### Improvements